added Increments observable that records the norm of displacements of a set of particles (to be tested). Also used another function to obtain unitary gaussian numbers.

+++ revreaddyC/src/observables/MeanSquaredDisplacement.cpp
@@ -37,7 +37,7 @@ void MeanSquaredDisplacement::record(World * world, double t) {
 	std::vector<double> squaredDisplacements;
 	double squaredDisplacement = 0.;
 	double mean = 0.;
-	int index = 0;
+	long index = 0;
 	unsigned int stillExisting = 0;
 	for (auto i=0; i < this->observedParticleIds.size(); i++) {
 		// check if particle still exists. if not: index = -1
